Great work on this package!

I have a simple trick that gives 10-20% faster encodes.

Avoids having to re-count individual bytes when match ends.

```
benchmark                     old ns/op     new ns/op     delta
BenchmarkCompress-8           4139335       3146599       -23.98%
BenchmarkCompressHC-8         31650000      30165988      -4.69%

benchmark                     old bytes     new bytes     delta
BenchmarkCompress-8           0             0             +0.00%
BenchmarkCompressHC-8         0             0             +0.00%

Before:
file	out	level	insize	outsize	millis	mb/s
enwik9	lz4	0	1000000000	489160485	6968	136.86
enwik9	lz4	1	1000000000	472338434	10406	91.65
enwik9	lz4	2	1000000000	405391862	20422	46.70

After:
file	out	level	insize	outsize	millis	mb/s
enwik9	lz4	0	1000000000	489160431	5342	178.52
enwik9	lz4	1	1000000000	472338287	9259	103.00
enwik9	lz4	2	1000000000	405391713	18551	51.41

Before:
file	out	level	insize	outsize	millis	mb/s
10gb.tar	lz4	0	10065157632	6441120237	71134	134.94
10gb.tar	lz4	1	10065157632	6369810447	117489	81.70
10gb.tar	lz4	2	10065157632	5697686196	266037	36.08

After:
file	out	level	insize	outsize	millis	mb/s
10gb.tar	lz4	0	10065157632	6441120013	65363	146.85
10gb.tar	lz4	1	10065157632	6369810471	110032	87.24
10gb.tar	lz4	2	10065157632	5697685530	244089	39.33

Before:
file	out	level	insize	outsize	millis	mb/s
adresser.json	lz4	0	7983034785	481827641	11168	681.70
adresser.json	lz4	1	7983034785	522398797	31439	242.16
adresser.json	lz4	2	7983034785	380853678	36492	208.62

After:
file	out	level	insize	outsize	millis	mb/s
adresser.json	lz4	0	7983034785	481827462	10197	746.54
adresser.json	lz4	1	7983034785	522398669	30017	253.63
adresser.json	lz4	2	7983034785	380853598	34770	218.95
```
It seems like the allocations are hiding the performance  of some of the benchmarks, so I did some tests on some bigger test data to make sure the performance was 'real'.

Edit: Yes, confirmed it in my other PR - the benchmarks don't show anything useful. Removed them above.

There is a few bytes of difference in size, which I can't really explain, if your bounds are correct - which they appear to be. Maybe do some additional tests.