@pierrec 2019-07-01
Wow nice trick, thanks Klaus.